Refactored to integrate generic optimisation and exporting parameters
